update rename attachment dialog
