+++ toolkit/featureforms/src/main/java/com/arcgismaps/toolkit/featureforms/internal/components/attachment/AttachmentTile.kt
@@ -431,88 +431,6 @@ private fun Size(
     )
 }
 
-@Composable
-internal fun RenameAttachmentDialog(
-    name: String,
-    onRename: (String) -> Unit,
-    onDismissRequest: () -> Unit
-) {
-    val groups = rememberSaveable(name) { name.split("\\.(?=[^\\\\.]+\$)".toRegex()) }
-    var filename by rememberSaveable(groups) { mutableStateOf(groups.first()) }
-    val extension = rememberSaveable(groups) { if (groups.count() == 2) groups.last() else "" }
-    val focusRequester = remember { FocusRequester() }
-    Dialog(
-        onDismissRequest = onDismissRequest,
-        properties = DialogProperties(usePlatformDefaultWidth = true),
-        content = {
-            Surface(
-                modifier = Modifier.wrapContentSize(),
-                shape = RoundedCornerShape(5.dp)
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = stringResource(R.string.rename_attachment),
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Bold
-                    )
-                    Spacer(modifier = Modifier.height(25.dp))
-                    TextField(
-                        value = TextFieldValue(
-                            text = filename,
-                            selection = TextRange(filename.length)
-                        ),
-                        onValueChange = { value -> filename = value.text },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .focusRequester(focusRequester)
-                            .onGloballyPositioned {
-                                focusRequester.requestFocus()
-                            },
-                        label = { Text(stringResource(R.string.name)) },
-                        suffix = {
-                            Text(text = ".$extension")
-                        },
-                        singleLine = true,
-                        keyboardOptions = KeyboardOptions.Default.copy(
-                            imeAction = ImeAction.Done,
-                        )
-                    )
-                    Spacer(modifier = Modifier.height(25.dp))
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceEvenly,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Button(onClick = onDismissRequest) {
-                            Text(text = stringResource(id = R.string.cancel))
-                        }
-                        Button(
-                            onClick = { onRename("$filename.$extension") },
-                            enabled = filename.isNotEmpty()
-                        ) {
-                            Text(text = stringResource(id = R.string.rename))
-                        }
-                    }
-                }
-            }
-        }
-    )
-}
-
-@Preview
-@Composable
-private fun RenameAttachmentDialogPreview() {
-    RenameAttachmentDialog(
-        name = "Photo 1.jpg",
-        onRename = {},
-        onDismissRequest = {}
-    )
-}
-
 @Preview
 @Composable
 private fun AttachmentTilePreview() {
